+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # ✒ CHANGELOG
 
+
+#### Version: 1.2.1 > 1.2.2
+``Date 17/11/2020 15:12``
+* Fix Logout Permissions for Members/Admins Roles
+* Fix menuactive undefined variable in Sidebar.volt (Menu active feature disabled temporary)
+* Fix Admin Permissions to access to member dashboard
+* Fix Password Profile requirement ( set to current password if no change posted )
+* Feature : Image uploader for models
+* Feature : Generating Upload path method ( Line ``Sakura\Models\ModelBase.php:309`` )
+
+
 #### Version: 1.2.0-alpha > 1.2.1  
 ``Date 02/11/2020 15:12``